changed float to int(sorry im dyslexic?)

+++ Movie Recommendation/Models/MResults.cs
@@ -30,6 +30,6 @@ namespace Movie_Recommendation.Models
 
 		public bool video { get; set; }
 
-		public int vote_average { get; set; }
+		public float vote_average { get; set; }
 	}
 }
